requirements.txt
amqp==5.1.1
aniso8601==9.0.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
decorator==5.1.1
executing==1.2.0
Flask==2.3.2
Flask-JWT-Extended==4.5.2
Flask-MySQL==1.5.2
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.3
flower==1.2.0
greenlet==2.0.2
humanize==4.6.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
kombu==5.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pika==1.3.2
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PyJWT==2.7.0
PyMySQL==1.0.3
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.15
stack-data==0.6.2
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.6.2
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.3.4
